Fix #1964 - Breakpoints not working in view (html.erb) files
The addition of JS debugging overtook setting breakpoints on ERB files. This \"fix\" is to specifically force ruby breakpoint setting for HTML/ERB files.

The longer term fix is to actually have the JS/Ruby debug code check the line in question and see if it's a valid one for a JS/Ruby breakpoint. (Because right now you can't do JS debugging inside HTML/ERB because of this fix).