#1907 - Parse ruby files using 1.8 vs 1.9 parser based on interpreter on PATH for the project
Allow passing down the CompatVersion to specify what source parser to use. BOTH or RUBY_19 uses the Ruby 1.9 parser. RUBY_18 uses the Ruby 1.8 parser. We now default ot using the BOTH (and therefore Ruby 1.9) unless overriden.

The only place it seems we really need to care about syntax differences/errors is in our ruby validator, so we sniff out the ruby interpreter version there and set the CompatVersion to match.